Create snoopy-pro.xml
Origional file needed if you want to go throw the steps the author made for writing the driver. This file is the reverse engineering of the windows driver. The owner of this file is not me, it is the author and I hope he include the file with the other files. Thanks /Mariwan